Added ResetPassword functionality; Refactored files to be in new project structure
