+++ src/components/ResetPassword/ResetPassword.js
@@ -1,17 +1,18 @@
 import React from "react";
-import { SignUp } from "aws-amplify-react";
+import { ForgotPassword } from "aws-amplify-react";
 
-export class CreateAccount extends SignUp {
+export class ResetPassword extends ForgotPassword {
   constructor(props) {
     super(props);
-    this._validAuthStates = ["signUp"];
+    this._validAuthStates = ["forgotPassword"];
   }
 
   showComponent() {
     return (
       <div>
-        <h1>Create a new account</h1>
+        <h1>Reset your password</h1>
         <form>
+          <h4>1. Receive a confirmation code</h4>
           <div>
             <label htmlFor="username">Email</label>
             <input
@@ -24,7 +25,29 @@ export class CreateAccount extends SignUp {
             />
           </div>
           <div>
-            <label htmlFor="password">Password</label>
+            <button
+              type="button"
+              onClick={() => {
+                super.send();
+              }}
+            >
+              Send Code
+            </button>
+          </div>
+          <h4>2. Use that confirmation code to change your password</h4>
+          <div>
+            <label htmlFor="code">Code * </label>
+            <input
+              id="code"
+              key="code"
+              name="code"
+              onChange={this.handleInputChange}
+              placeholder="Enter your code"
+              autoComplete="off"
+            />
+          </div>
+          <div>
+            <label htmlFor="password">New Password * </label>
             <input
               id="password"
               key="password"
@@ -32,21 +55,20 @@ export class CreateAccount extends SignUp {
               onChange={this.handleInputChange}
               type="password"
               placeholder="******************"
+              autoComplete="off"
             />
           </div>
           <div>
             <button
               type="button"
               onClick={() => {
-                super.sortFields();
-                super.validate();
-                super.signUp();
+                super.submit();
               }}
             >
-              Create Account
+              Submit
             </button>
             <p>
-              Have an account?{" "}
+              Back to{" "}
               <span onClick={() => super.changeState("signIn")}>Sign In</span>
             </p>
           </div>
